pyproject.toml
[project]
name = "distributed-data-analytics-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.25",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.7",
    "openai>=1.86.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "streamlit>=1.45.1",
]
